implement eventual consistency via a middleware

+++ src/GymManagement.Application/Subscriptions/Events/SubscriptionDeletedEventHandler.cs
@@ -14,8 +14,13 @@ public class SubscriptionDeletedEventHandler(
 
     public async Task Handle(SubscriptionDeletedEvent notification, CancellationToken cancellationToken)
     {
-        var subscription = await _subscriptionRepository.GetByIdAsync(notification.SubscriptionId)
-            ?? throw new InvalidOperationException();
+        var subscription = await _subscriptionRepository.GetByIdAsync(notification.SubscriptionId);
+
+        if (subscription is null)
+        {
+            //resilient error handling 
+            throw new InvalidOperationException();
+        }
 
         await _subscriptionRepository.RemoveSubscriptionAsync(subscription);
         await _unitOfWork.CommitChangesAsync();
